Updates register page with new updates
